pdf_reader_llm_app/requirements.txt
# Core Flask framework
Flask==2.3.3
Werkzeug==2.3.7

# LangChain and related packages
langchain==0.1.0
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langchain-core==0.1.10
langchain-nomic==0.0.1

# Document processing
PyPDF2==3.0.1
pypdf==3.17.4

# Vector store
chromadb==0.4.18

# Environment variables
python-dotenv==1.0.0

# OpenAI API (for LLM)
openai==1.6.1

# Additional dependencies that might be needed
numpy==1.24.3
tiktoken==0.5.2
tenacity==8.2.3
requests==2.31.0
urllib3==2.0.7
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# For embeddings
sentence-transformers==2.2.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Optional: for better PDF processing
pdfplumber==0.10.0